hw16/src/main/java/ru/otus/hw/rest/GenreRestController.java
package ru.otus.hw.rest;

import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;
import ru.otus.hw.dto.GenreDto;
import ru.otus.hw.services.GenreService;

import java.util.List;

import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;

@RestController
@RequiredArgsConstructor
public class GenreRestController {

    private final GenreService genreService;

    @GetMapping(value = "api/v1/genres", produces = APPLICATION_JSON_VALUE)
    public List<GenreDto> getAllGenre() {
        return genreService.findAll();
    }
}
